// swiyu-trust-management-ui/src/app/api/generated/model/paged-model-trust-onboarding-submission-document-list-item-dto.ts
/**
 * SWIYU Trust Management SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {PageMetadata} from './page-metadata';
import {TrustOnboardingSubmissionDocumentListItemDto} from './trust-onboarding-submission-document-list-item-dto';

export interface PagedModelTrustOnboardingSubmissionDocumentListItemDto {
  content?: Array<TrustOnboardingSubmissionDocumentListItemDto>;
  page?: PageMetadata;
}