// swiyu-trust-management-ui/src/app/api/generated/model/trust-onboarding-task-action.ts
/**
 * SWIYU Trust Management SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export const TrustOnboardingTaskAction = {
  Reject: 'REJECT',
  Approve: 'APPROVE',
  RequestMoreInformation: 'REQUEST_MORE_INFORMATION',
  AddInternalNote: 'ADD_INTERNAL_NOTE',
  AssignSelf: 'ASSIGN_SELF'
} as const;
export type TrustOnboardingTaskAction = (typeof TrustOnboardingTaskAction)[keyof typeof TrustOnboardingTaskAction];