// swiyu-trust-management-ui/src/app/api/generated/model/trust-onboarding-task-status.ts
/**
 * SWIYU Trust Management SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export const TrustOnboardingTaskStatus = {
  Rejected: 'REJECTED',
  Accepted: 'ACCEPTED',
  Opened: 'OPENED',
  InformationRequested: 'INFORMATION_REQUESTED'
} as const;
export type TrustOnboardingTaskStatus = (typeof TrustOnboardingTaskStatus)[keyof typeof TrustOnboardingTaskStatus];